Improve queue reliability and concurrency
Replaces Bee Queue in favor of using a Redis Streams
backed approach. Workers will join a consumer
group on a Redis Stream and pull tasks.

In the case a task fails, it is not acknowledged
and the error is logged. The task will be
retried indefinitely.

Furthermore, this commit also improves the
concurrency model of the worker using worker threads,
where the old approach (ab)used the event loop.

This refactor *will* need a refactor itself
to clean the code up considerably.